Consolidated all shared database code for H2 and PostgreSQL
