+++ src/main/java/com/timmattison/cryptocurrency/modules/BitcoinNoLoopTestModule.java
@@ -13,7 +13,7 @@ import com.timmattison.cryptocurrency.bitcoin.*;
 import com.timmattison.cryptocurrency.bitcoin.factories.*;
 import com.timmattison.cryptocurrency.factories.*;
 import com.timmattison.cryptocurrency.interfaces.*;
-import com.timmattison.cryptocurrency.standard.H2BlockStorage;
+import com.timmattison.cryptocurrency.standard.blockstorage.database.H2BlockStorage;
 import com.timmattison.cryptocurrency.standard.NoLoopVariableLengthInteger;
 import com.timmattison.cryptocurrency.standard.StandardBlockFactory;
 import com.timmattison.cryptocurrency.standard.StandardMerkleRootCalculator;
